const int gravity = 9.81

+++ vesc_driver/include/vesc_driver/vesc_driver.hpp
@@ -90,7 +90,7 @@ private:
   CommandLimit position_limit_;
   CommandLimit servo_limit_;
   std::string imu_frame_;
-  const int GRAVITATIONAL_ACC = 9.80665;
+  const double GRAVITATIONAL_ACC = 9.80665;
 
   // ROS services
   rclcpp::Publisher<VescStateStamped>::SharedPtr state_pub_;
